Add opentracing server-side tags
Add opentracing-defined server-side tags to spans created by this
middleware. These tags add information regarding the origin of the
span and also denote that the span was created by a server. This kind of
information can later be used by the tracer (e.g. the jaeger-client adds
Zipkin compatible 'ss', 'sr' annotations based on this information).