Add support for tracing hooks
Add hooks to be called when a trace starts or finishes. This acts as an
integration point to interact with the spans handled by this Django
middleware.

The hooks are python callables and they are provided by the Django
settings in the form of a dictionary:

  {'start': start_hook, 'finish': finish_hook}

The start hook expects the newly created span as argument.
The finish_hook expects the span that is ready to be finished.

Any of them can be omitted.