Do not handle the span if a hook decides so
Add support to defer managing the spans to an external entity decided by
the provided hooks. A hook can return a 'True' value and let the
middleware know that it should not handle the Span.

- If the 'start_hook' returns true, then the middleware should not keep
  any reference to the Span.

- If the 'finish_hook' returns true, then no further handling is needed
  for the span in hands.
  Note that based on the return value of the start hook, a reference to
  a Span for the given request may not even exist, but the finish hook
  must be called with a `None` argument.

This allows of better integration with a Span manager and enables
advanced scenarios (e.g. the Span manager accumulates and report spans
in batches for performance reasons).